requirements.txt
pyyaml
Django==4.2.7
psycopg2-binary==2.9.9
djangorestframework==3.14.0
